+++ CHANGES.rst
@@ -1,3 +1,9 @@
+Version 1.8.1
+-------------
+
+Unreleased
+
+
 Version 1.8.0
 -------------